+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Instant VRC Light Volumes V2 setup and support
+
 ### Fixed
 - Enabling lock state while playing won't lock progress bar ([#71](https://github.com/JLChnToZ/VVMW/issues/71))
 - Audio controller unable to function